wishlist added

+++ WalkInStyleAPI/Models/CartItem.cs
@@ -11,7 +11,7 @@ namespace WalkInStyleAPI.Models
         public int ProductId {  get; set; }
         public Product Product { get; set; }
         public int Quantity { get; set; }
-        public int TotalPrice { get; set; }
+        public decimal TotalPrice { get; set; }
 
     }
 }
