+++ WalkInStyleAPI/Models/DTOs/Cart/CartViewDto.cs
@@ -5,8 +5,8 @@
         public int ProductId { get; set; }
         public string? ProductName { get; set; }
         public string? Image { get; set; }
-        public int Price { get; set; }  
+        public decimal Price { get; set; }  
         public int Quantity { get; set; }
-        public int TotalPrice { get; set; }
+        public decimal TotalPrice { get; set; }
     }
 }
