+++ WalkInStyleAPI/Models/DTOs/Product/ProductDto.cs
@@ -4,8 +4,8 @@
     {
         public string? ProductName { get; set; }
         public string? Image { get; set; }
-        public int OrginalPrice { get; set; }
-        public int OfferPrice { get; set; }
+        public decimal OrginalPrice { get; set; }
+        public decimal OfferPrice { get; set; }
         public string? Description { get; set; }
         public string? Brand { get; set; }
         public int CategoryId { get; set; }
