+++ WalkInStyleAPI/Models/DTOs/Product/ProductViewDto.cs
@@ -5,8 +5,8 @@
         public int ProductId { get; set; }
         public string? ProductName { get; set; }
         public string? Image { get; set; }
-        public int OrginalPrice { get; set; }
-        public int OfferPrice { get; set; }
+        public decimal OrginalPrice { get; set; }
+        public decimal OfferPrice { get; set; }
         public string? Description { get; set; }
         public string? Brand { get; set; }
         public int CategoryId { get; set; }
